requirements.txt
PySide6==6.6.1
SQLAlchemy==2.0.25
cryptography==42.0.2
python-dotenv==1.0.0
Pillow==10.2.0
opencv-python==4.8.1.78 